Allow empty vmaas provider config block.
The way that client.GetServiceSettingsMap() works is unfortunately a bit
misleading and we probably need to change the return values from this
function.  Since when that function is run in NewClient we cannot know
if the customer is going to use VMaaS we need to return nil and no
error.  If the stanza is missing it means that the client won't be
initialised which will be picked up by GetClientFromMetaMap() which
returns an error to check that vmaas block is set.

We need to do this so that customers can decide which GL services they
want to use and supply the provider config blocks only for those
services.