Remove iam_token from client.go
In this PR we remove iam_token from client.go.  This field is no longer
available from the hpegl provider and has been replaced with
programmatic generation of tokens.  The first implementation of this
uses service-client creds. Since the token is required as a temporary
workaround for acceptance tests we get its value from the
HPEGL_IAM_TOKEN env-var.