Add API_URL as schema parameter
The standard pattern for the hpegl Terraform provider is to
allow the API endpoint of a service to be configurable as a parameter
in the service section of the provider block of a Terraform config.
This changes brings the VMaaS provider in line with that pattern.
A sample Terraform config would now look like:

    provider "hpegl" {
      vmaas {
        location   = var.location
        space_name = var.space
        api_url = "https://iac-vmaas.intg.hpedevops.net"
      }
    }

If not provided the api_url will default to the VMaaS production endpoint.