Fix-up docs templates
We've updated the version of tfplugin-docs which has exposed issues with
the existing resource templates.  As a result of a change in indentation
in the generated "description" field of the the .md files resulting from
the resource templates the documentation wasn't being categorised
correctly on the provider registry landing page.  To fix this we make
the following change in each template:

  {{ .Description | plainmarkdown | trimspace | prefixlines "  " }}

becomes

{{ .Description | plainmarkdown | trimspace | prefixlines "  " }}