Reuse token retrieve function in BrokerClient
The MetaFunction used to retrieve IAM API Client tokens in the
BrokerClient created when initialising the Client creates a new Token
Retrieve Function and a Token Handler each time that it is called. This
means that an IAM call is made on every call and there is no IAM Token
reuse.  We fix this so that we create just the one handler and token
retrive function and reuse in the MetaFunction.